requirements.txt
streamlit
streamlit-option-menu
seaborn
statsmodel
sklearn
matplotlib
plotly